Refactor project for Python package distribution
- Restructure project into proper Python package layout with src/ directory
- Split monolithic pdf_to_audio.py into logical modules (cli, core, api, image, utils, constants)
- Migrate from Poetry to modern PEP 621 compliant pyproject.toml with Hatchling
- Add comprehensive test suite with pytest
- Update README with detailed installation and usage instructions
- Add MANIFEST.in for proper package distribution
- Enable installation via pip and pipx as CLI tool
- Maintain backward compatibility with existing CLI interface
- Add development dependencies and tooling configuration
- Create distributable wheel and source packages
