+++ REFACTORING_SUMMARY.md
@@ -1,0 +1,182 @@
+# PDF to Audio - Refactoring Summary
+
+## Overview
+
+This document summarizes the refactoring of the `pdf-to-audio` project to make it distributable as a Python package and installable CLI tool via pipx.
+
+## Changes Made
+
+### 1. Project Structure Reorganization
+
+**Before:**
+```
+pdf-to-audio/
+├── pdf_to_audio.py  (monolithic file ~570 lines)
+├── pyproject.toml   (mixed Poetry/PEP 621 syntax)
+├── poetry.lock
+├── README.md        (minimal)
+└── LICENSE
+```
+
+**After:**
+```
+pdf-to-audio/
+├── src/
+│   └── pdf_to_audio/
+│       ├── __init__.py      # Package initialization and version info
+│       ├── cli.py           # Command-line interface and argument parsing
+│       ├── core.py          # Core document processing logic
+│       ├── api.py           # Mistral API interaction functions
+│       ├── image.py         # Image processing and description functions
+│       ├── utils.py         # Utility functions (tokens, chunking, etc.)
+│       └── constants.py     # Constants and system prompts
+├── tests/
+│   ├── __init__.py
+│   ├── test_cli.py          # CLI functionality tests
+│   └── test_utils.py        # Utility function tests
+├── dist/                    # Built packages (wheel and sdist)
+├── pyproject.toml           # Modern PEP 621 configuration
+├── MANIFEST.in              # Package manifest
+├── README.md                # Comprehensive documentation
+└── LICENSE
+```
+
+### 2. Build System Migration
+
+- **Removed:** Poetry-specific configuration and `poetry.lock`
+- **Added:** Modern PEP 621 compliant `pyproject.toml` using Hatchling as build backend
+- **Benefits:** 
+  - Standard Python packaging
+  - Compatible with all Python package managers (pip, pipx, conda, etc.)
+  - Faster builds and smaller dependencies
+
+### 3. Package Configuration (`pyproject.toml`)
+
+**Key improvements:**
+- Proper PEP 621 metadata format
+- Comprehensive package classifiers
+- Optional dependencies for development and interactive use
+- Correct entry point configuration for CLI
+- Minimum Python version lowered to 3.8 for broader compatibility
+
+### 4. Code Modularization
+
+The monolithic `pdf_to_audio.py` file was split into logical modules:
+
+- **`cli.py`**: Command-line interface, argument parsing, and main entry point
+- **`core.py`**: Document processing workflow and orchestration
+- **`api.py`**: Mistral API interactions with retry logic
+- **`image.py`**: Image processing and description generation
+- **`utils.py`**: Utility functions for text processing and chunking
+- **`constants.py`**: Configuration constants and system prompts
+
+### 5. Enhanced Documentation
+
+- **README.md**: Complete rewrite with:
+  - Installation instructions for pip and pipx
+  - Comprehensive usage examples
+  - Command-line options documentation
+  - Development setup instructions
+  - Mathematical notation conversion examples
+
+### 6. Testing Infrastructure
+
+- Added pytest-based test suite
+- Unit tests for utility functions
+- CLI functionality tests
+- Test coverage for core functionality
+- Development dependencies for testing and code quality
+
+### 7. Distribution Readiness
+
+- **Wheel and source distributions** can be built with `python -m build`
+- **pipx compatible**: Can be installed as isolated CLI tool
+- **pip installable**: Standard Python package installation
+- **Development mode**: Supports `pip install -e .` for development
+
+## Installation Methods
+
+### End Users
+
+```bash
+# Via pipx (recommended for CLI tools)
+pipx install pdf-to-audio
+
+# Via pip
+pip install pdf-to-audio
+
+# From source
+git clone https://github.com/sjvrensburg/pdf-to-audio.git
+cd pdf-to-audio
+pip install .
+```
+
+### Developers
+
+```bash
+git clone https://github.com/sjvrensburg/pdf-to-audio.git
+cd pdf-to-audio
+pip install -e ".[dev]"
+```
+
+## Quality Assurance
+
+### Testing
+```bash
+pytest                    # Run tests
+pytest --cov             # Run with coverage
+```
+
+### Code Quality
+```bash
+ruff check .              # Linting
+ruff format .             # Code formatting
+isort .                   # Import sorting
+mypy .                    # Type checking
+```
+
+### Building
+```bash
+python -m build           # Build wheel and sdist
+```
+
+## Benefits of Refactoring
+
+1. **Maintainability**: Modular code structure makes it easier to maintain and extend
+2. **Testability**: Separated concerns allow for comprehensive unit testing
+3. **Distributability**: Standard Python packaging enables easy distribution
+4. **CLI Tool**: Can be installed as a standalone CLI tool via pipx
+5. **Development**: Better development experience with proper tooling support
+6. **Documentation**: Comprehensive documentation for users and developers
+7. **Standards Compliance**: Follows modern Python packaging standards (PEP 621)
+
+## Backward Compatibility
+
+The CLI interface remains exactly the same, so existing users can upgrade seamlessly:
+
+```bash
+pdf-to-audio input.pdf output.txt --include_images --verbose
+```
+
+All command-line options and functionality are preserved.
+
+## Future Enhancements
+
+The modular structure now makes it easy to add:
+- Additional output formats
+- New AI model providers
+- Plugin system for custom processing
+- Web interface
+- API server mode
+- Configuration file support
+
+## Verification
+
+The refactored package has been tested and verified to:
+- ✅ Install via pip and pipx
+- ✅ Build wheel and source distributions
+- ✅ Pass all unit tests
+- ✅ Maintain CLI compatibility
+- ✅ Import correctly as a Python package
+- ✅ Handle missing API keys gracefully
+- ✅ Support all original command-line options
