+++ src/pdf_to_audio/__init__.py
@@ -1,0 +1,9 @@
+"""PDF to Audio - Convert PDFs to TTS-friendly text using Mistral API."""
+
+__version__ = "0.1.0"
+__author__ = "sjvrensburg"
+__email__ = "stefanj@mandela.ac.za"
+
+from .cli import main
+
+__all__ = ["main"]
